회원 컨트롤러를 분리 합니다.
